ViewNote
ViewNote and Create note sections loading when link is clicked
