+++ lambda-notes/src/components/MainPage/ViewNotes.js
@@ -2,19 +2,12 @@ import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import { Route, Link } from 'react-router-dom';
 import { fetchingNotes } from '../../actions/index';
+import SingleNote from './Note'
 import styled from 'styled-components';
 
 const Header = styled.div`
 display: flex;
 width: 100%;
-`
-
-const NoteContainer = styled.div`
-margin 20px;
-border: 1px solid black;
-background-color: white;
-width: 200px;
-height: 150px;
 `
 
 const Main = styled.div`
@@ -34,23 +27,11 @@ class ViewNotes extends Component {
             < Main >
                 <Header><h1>Your Notes:</h1></Header>
                 {this.props.notes.map(note => (
-                    <SingleNote key={note.title} note={note} />
+                    <Link to='view-note'> <SingleNote key={note._id} note={note} /></Link>
                 ))}
             </Main >
         );
     }
-}
-
-function SingleNote({ note }) {
-    const { tags, title, textBody } = note;
-    return (
-        <Link to='view-note'>
-            <NoteContainer>
-                <div className='headline'>{note.title}</div>
-                <div className='content'>{note.textBody}</div>
-            </NoteContainer>
-        </Link>
-    );
 }
 
 const mapStateToProps = state => {
